+++ packages/core/rpc_dart_opentelemetry/lib/rpc_dart_opentelemetry.dart
@@ -5,5 +5,8 @@
 library;
 
 export 'src/interceptor/otel_rpc_interceptor.dart';
+export 'src/interceptor/otel_rpc_client_interceptor.dart';
 export 'src/propagation/rpc_otel_propagator.dart';
 export 'src/metrics/rpc_otel_metrics.dart';
+export 'src/metrics/rpc_status_names.dart';
+export 'src/bridge/log_controller_otel_output.dart';